23-cc || fixes sprint 2 and added endpoint for search transactions by amount

+++ microservices/ms-accounts/src/main/java/com/digitalmoney/msaccounts/persistency/dto/TransactionResponseDTO.java
@@ -12,7 +12,7 @@ public record TransactionResponseDTO (Long accountId,
                                       LocalDateTime transactionDate,
                                       String transactionDescription,
                                       String destinationCvu,
-                                      Integer transactionId,
+                                      Long transactionId,
                                       String originCvu,
                                       Transaction.TransactionType transactionType
 ) {
